refactor: reoder files and modified paths component imports
