+++ app/contact/components/Faq/Faq.tsx
@@ -1,10 +1,11 @@
 import { ListItem } from './components';
-import { faq } from '../../data/faq';
+import { FAQ } from './components/ListItem/data';
+
 export const Faq = () => {
   return (
     <div>
       <h1>¿Tienes alguna pregunta?</h1>
-      <ListItem questionList={faq} />
+      <ListItem questionList={FAQ} />
     </div>
   );
 };
