+++ app/contact/page.tsx
@@ -1,3 +1,5 @@
+import { NavBar } from '@app/components';
+
 import { Card, Faq, Form } from './components';
 import styles from './styles/contact.module.css';
 
@@ -7,18 +9,21 @@ export const metadata = {
 };
 
 const Contact = () => (
-  <div className={styles.container}>
-    <section className={styles['grid-contact']}>
-      <div className={styles['grid-contact__form']}>
-        <Form />
-      </div>
-      <div className={styles['grid-contact__social']}>
-        <Card />
-      </div>
-    </section>
-    <section className={styles['section-question']}>
-      <Faq />
-    </section>
+  <div className={styles['main-container']}>
+    <NavBar />
+    <div className={styles['container-contact']}>
+      <section className={styles['grid-contact']}>
+        <div className={styles['grid-contact__form']}>
+          <Form />
+        </div>
+        <div className={styles['grid-contact__social']}>
+          <Card />
+        </div>
+      </section>
+      <section className={styles['section-question']}>
+        <Faq />
+      </section>
+    </div>
   </div>
 );
